Handle /data/graph requests using real data from the repository
Run a SPARQL construct query extracting an interesting subgraph
around the queried-for nodes, ie. these nodes with their types and names
(FOAF names or DC titles) plus directly connected nodes with their
respective types and names (this omits untyped or unnamed nodes).

Process the resulting triples into links and nodes data to return to the
client.

The process is rather simplistic at the moment. In particular it:
- does not put "favourite" on links,
- just uses type/predicate IRIs as "type" values - which the sample
frontend does not understand, so everything is going to look gray,
- handles multiple names for a single node nondeterministically.