Support remote SPARQL Protocol repositories (eg. Blazegraph)
Create third Spring profile - "sparql".
In this profile set up a SPARQL repository using the
"comac.repository-url" configuration property to specify the repository
location.
This single-URL configuration will not work for updates on repositories
where the update URL is different from the query URL (such as remote
Sesame repositories, but we already have another profile for Sesame).

Add dependency on the sesame-repository-sparql module (it is currently
also pulled in by the memory sail for some reason, but let's not count
on it).

Do not yet change the default profile in application.properties.