Use Blazegraph integrated full text search
The SPARQL query used for search used simple substring filtering,
which was rather slow (and did not care for word boundaries, etc.).
Change it to use the FTS feature instead. Require matching all search
terms to avoid thinking much about relevance at the moment.

The Blazegraph FTS extension is described at
https://wiki.blazegraph.com/wiki/index.php/FullTextSearch
and requires the used store to be created with text indexing enabled.